Timeline showing events + structural improvements and refactoring
Added timeline with blue dots for each rendered graph and gray dots for the rest of the graphs in the selected time span.
Added links to individual and aggregated views from /.
Removed unused javascript and css files.
Renamed AggregateGraphs to aggregate-graphs and IndividualGraphs to individual-graphs.
Added doc-strings to functions and comments to clarify the code.
Restructured the filter.js file to create a function that updates the new timeline and text.
Changed from Tracker.autorun to observeChanges which makes an update on each added graph.
Added polling of changes to prevent the browser from crashing.
Reformatted the css file.
Restructured the client main.js to handle single graphs instead of a list of graphs.
Replaced the Graphs object with a variable each for the collections.
Moved setColor to a new javascript file color in the lib folder.
Added new dependencies to the public folder for offline development.